Revert "Enhanced"
