+++ src/components/AddItem.jsx
@@ -1,10 +1,6 @@
 import React from 'react';
-import { useItems } from '../context/ItemsContext';
-import { useState } from 'react';
 
-export default function AddItem() {
-  const { handleAddItem } = useItems();
-  const [text, setText] = useState('');
+export default function AddItem({ handleAddItem, text, setText }) {
   return (
     <form>
       <label>
@@ -15,14 +11,7 @@ export default function AddItem() {
           onChange={(e) => setText(e.target.value)}
         ></input>
       </label>
-      <button
-        onClick={(e) => {
-          handleAddItem(e, text);
-          setText('');
-        }}
-      >
-        add item
-      </button>
+      <button onClick={(e) => handleAddItem(e, text)}>add item</button>
     </form>
   );
 }
